+++ packages/vsce/src/webviews/src/components/PanelView.tsx
@@ -1,0 +1,14 @@
+import { VSCodePanelView } from "@vscode/webview-ui-toolkit/react";
+
+interface ViewProperty {
+  id: string;
+  children: React.ReactNode;
+}
+
+export default function PanelView(prop: ViewProperty) {
+  return (
+    <VSCodePanelView id={prop.id} className="set-margin">
+      {prop.children}
+    </VSCodePanelView>
+  );
+}